+++ ads_mcp/coordinator.py
@@ -38,6 +38,7 @@ if _CLIENT_ID and _CLIENT_SECRET:
             "https://www.googleapis.com/auth/userinfo.email",
             "https://www.googleapis.com/auth/userinfo.profile",
             "https://www.googleapis.com/auth/adwords",
+            "https://www.googleapis.com/auth/webmasters",
         ],
     )